refactor(Account): use BigDecimal in place of double in balance attribute.

+++ src/main/java/com/billetera_virtual/account/domain/port/AccountServicePort.java
@@ -3,6 +3,7 @@ package com.billetera_virtual.account.domain.port;
 import com.billetera_virtual.account.domain.Account;
 import com.billetera_virtual.account.domain.dto.AccountPublicDataResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountServicePort {
@@ -13,11 +14,11 @@ public interface AccountServicePort {
     Account getAccountByCvu(String cvu);
     AccountPublicDataResponse getAccountPublicData(String identifier, Long accountId);
     List<Account> getAll();
-    Account updateAccount(Long id, double balance);
+    Account updateAccount(Long id, BigDecimal balance);
     void updateAlias(Long id, String alias);
     void logicallyDeleteById(Long id);
     Account createAccountFromUser(Long userId);
-    double makeTransaction(Long originId, Long counterpartyId, double amount);
+    BigDecimal makeTransaction(Long originId, Long counterpartyId, BigDecimal amount);
     Long getAccountIdByUserId(Long userId);
     Long getAccountIdByAlias(String alias);
     Long getAccountIdByDestination(String destination);
